+++ CHANGELOG.md
@@ -12,6 +12,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 ### Security
 
+## [0.2.7]
+### Added
+- `Mastrubate` option available via `self` menu
+  - only `finger ass` option available so far but more to be added later
+
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
 ## [0.2.6] - 2017-12-18
 ### Added
 - A message when entering an area for the first time